Marking current PowerShell configuration obsolete via rename
The issue is the way interaction with PowerShell in the System.Management.Automation package works. Perhaps I have a lot to learn to make it work, but writing to the console does not work when apps don't use the PowerShell sinks, I'm looking at you Scoop!

The second issue the System.Management.Automation package is compatibility with .NET 6. In order to move to .NET 6 I am choosing the path of less resistance and plan to drop the dependecy on System.Management.Automation all together.